[fix / feature] 로그인시 작동방식 변경
추후과제: nav작동을 login여부에 맞춰서 해야함
